Make a quick example of switchable cookie env vars
